cli/login: move safe_format_key and its tests from codex_login to cli module
